게시글 첨부 파일 불러오기 완료. 수정 진행중

+++ src/main/java/com/sbs/cyj/at/dao/ArticleDao.java
@@ -17,7 +17,7 @@ public interface ArticleDao {
 
 	public void write(Map<String, Object> param);
 
-	Article getArticleById(long id);
+	Article getArticleById(int id);
 
 	void modify(Map<String, Object> param);
 }
